.sql file, the migration.js file within the API directory, users.js, and products.js. The changes were driven primarily by a need to improve data handling across these files.
In "migrator.js", more descriptive error handling was added and migrated detailing from the "E01_FACTURA" to the "E01_DETALLE_FACTURA" collection. This change offers a more effective snapshot of migrated details.

In the "users.js" file, specific error messages and additional input validation were introduced to provide users with more precise feedback. A user's ability to securely interact with their data increases with these additions.

In "product.js", connection error handling was improved and more detailed responses were added, providing users better and appropriate feedback.

In addition, unnecessary table creation in the "E01_DETALLE_FACTURA" was removed from ITBA_2023_esquema_facturacion.sql, which streamlined the schema.

Lastly, formatting adjustments were made in multiple areas to ensure code readability and documentation consistency.

+++ API/mongodb/users.js
@@ -1,17 +1,24 @@
 const express = require('express');
 const router = express.Router();
-const {getdb}= require('./datasource')
+const {getdb} = require('./datasource')
 
 //todo error handling
 
 //todo api methods
 
 router.get('/', async (req, res) => {
+    let db = null;
     try {
-        const db = await getdb();
+        db = await getdb();
+    } catch (err) {
+        console.error("DB connection error:", err);
+        return res.status(500).json({error: "DB connection error"}).end();
+    }
+
+    try {
         const result = await db.collection("E01_CLIENTE").find({}).toArray();
-    
-        if(result.length === 0) {
+
+        if (result.length === 0) {
             res.status(404).json({message: "No data found"});
         } else {
             res.status(200).json(result);
@@ -21,20 +28,55 @@ router.get('/', async (req, res) => {
     }
 });
 
-router.post('/', async  (req, res, next) => {
-   const {nro_cliente, nombre, apellido, direccion, activo} = req.body;
-   try {
-       const db = await getdb();
-       await db.collection("E01_CLIENTE").insertOne({nro_cliente, nombre, apellido, direccion, activo});
-       res.status(201).end();
-   } catch (err) {
-       res.status(500).end();
-   }
+router.post('/', async (req, res, next) => {
+    const {nro_cliente, nombre, apellido, direccion, activo, telefonos} = req.body;
+
+    if (!nro_cliente || !nombre || !apellido || !direccion || !activo) {
+        return res.status(400).json({error: 'Missing required fields. Please provide values for nro_cliente, nombre, apellido, direccion, and activo.'});
+    }
+
+    if (telefonos && Array.isArray(telefonos)) {
+        for (const telefono of telefonos) {
+            if (typeof telefono.codigo_area === 'undefined' ||
+                typeof telefono.nro_telefono === 'undefined' ||
+                typeof telefono.tipo === 'undefined') {
+                return res.status(400).json({
+                    error: 'Missing required fields for telefono. Please provide values for codigo_area, nro_telefono, and tipo.'
+                });
+            }
+        }
+    } else if (!Array.isArray(telefonos)) {
+        return res.status(400).json({
+            error: 'telefonos is expected to be an array'
+        });
+    }
+
+
+    try {
+        const db = await getdb();
+        await db.collection("E01_CLIENTE").insertOne({nro_cliente, nombre, apellido, direccion, activo, telefonos});
+        res.status(201).end();
+    } catch (err) {
+        if (err.code === 11000) {
+            // Duplicate key error, return a bad request
+            res.status(400).json({error: 'Duplicate key violation. The specified nro_cliente already exists.'});
+        } else {
+            console.error(err);
+            // Other errors, return internal server error
+            res.status(500).end();
+        }
+    }
 });
 
 router.get('/:id', async (req, res, next) => {
     const {id} = req.params;
-    const db = await getdb();
+    let db = null;
+    try {
+        db = await getdb();
+    } catch (err) {
+        console.error("DB connection error:", err);
+        return res.status(500).end();
+    }
 
     try {
         const result = await db.collection("E01_CLIENTE").findOne({nro_cliente: parseInt(id)})
@@ -49,19 +91,43 @@ router.get('/:id', async (req, res, next) => {
     }
 
 });
+
 router.delete('/:id', async (req, res) => {
     const {id} = req.params;
     const db = await getdb();
-    db.collection("E01_CLIENTE").deleteOne({nro_cliente: parseInt(id)});
-    res.status(200).end();
-});
 
+    // Validate Input
+    const isValid = Number.isInteger(id);
+
+    if (!isValid) {
+        return res.status(400).json({error: 'Invalid client ID.'});
+    }
+
+    try {
+        // Perform DB operation
+        const deletionResult = await db.collection("E01_CLIENTE").deleteOne({nro_cliente: parseInt(id)});
+
+        if (deletionResult.deletedCount === 0) {
+            return res.status(404).json({error: 'Client not found.'});
+        }
+
+        res.status(200).json({success: 'Client deleted successfully.'});
+    } catch (err) {
+        console.error(err);
+        res.status(500).json({error: 'An error occurred, please try again later.'});
+    }
+});
 router.put('/:id', async (req, res) => {
     const {id} = req.params;
-    const {nombre, apellido, direccion, activo} = req.body;
+
+    if (!id) {
+        return res.status(400).json({error: 'Missing required fields. Please provide values for id.'});
+    }
+
+    const {nombre, apellido, direccion, activo, telefonos} = req.body || {};
 
     const setData = {};
-    if ( nombre !== null && nombre !== undefined ){
+    if (nombre !== null && nombre !== undefined) {
         setData.nombre = nombre;
     }
 
@@ -77,11 +143,47 @@ router.put('/:id', async (req, res) => {
         setData.activo = activo
     }
 
-    const db = await getdb();
+    if (telefonos && Array.isArray(telefonos)) {
+        for (const telefono of telefonos) {
+            if (typeof telefono.codigo_area === 'undefined' ||
+                typeof telefono.nro_telefono === 'undefined' ||
+                typeof telefono.tipo === 'undefined') {
+                return res.status(400).json({
+                    error: 'Missing required fields for telefono. Please provide values for codigo_area, nro_telefono, and tipo.'
+                });
+            }
+        }
+        setData.telefonos = telefonos;
+    } else {
+        return res.status(400).json({
+            error: 'telefonos is expected to be an array'
+        });
+    }
 
-    const result = await db.collection("E01_CLIENTE").updateOne({nro_cliente: parseInt(id)}, { $set : setData});
+    let db = null;
+    try {
+        db = await getdb();
+    } catch (err) {
+        console.error("DB connection error:", err);
+        return res.status(500).end();
+    }
 
-    result.modifiedCount === 1 ? res.status(200).end() : res.status(404).end();
+    let result = null;
+
+    try {
+        result = await db.collection("E01_CLIENTE").updateOne({nro_cliente: parseInt(id)}, {$set: setData});
+    } catch (err) {
+        if (err.code === 11000) {
+            // Duplicate key error, return a bad request
+            return res.status(400).json({error: 'Duplicate key violation.'});
+        } else {
+            console.error(err);
+            // Other errors, return internal server error
+            return res.status(500).end();
+        }
+    }
+
+    result.modifiedCount === 1 ? res.status(200).end() : res.status(404).json({error: "Record not found"}).end();
 
 });
